Moved dummy model to its own file
