+++ test/unit/test_model_flavor.py
@@ -248,25 +248,13 @@ class TestModelFlavors:
         from sklearn.datasets import load_iris
         from sklearn.svm import SVC
 
-        from layer import CustomModel
-
-        class DummyModel(CustomModel):
-            def __init__(self):
-                super().__init__()
-                svc = SVC()
-                svc.set_params(kernel="linear")
-                x, y = load_iris(return_X_y=True)
-                svc.fit(x, y)
-                self.model = svc
-
-            def predict(self, model_input):
-                return self.model.predict(model_input)
+        from .common.dummy_model import DummyModel
 
         model = DummyModel()
         flavor = get_flavor_for_model(model)
         assert type(flavor).__name__ == CustomModelFlavor.__name__
 
-        flavor.save_model_to_directory(DummyModel(), tmp_path)
+        flavor.save_model_to_directory(model, tmp_path)
         loaded_model = flavor.load_model_from_directory(tmp_path).model_object
         assert isinstance(loaded_model, layer.CustomModel)
         assert isinstance(loaded_model.model, SVC)
